refactor(react-gui): give the scene clipboard a folder
`sceneClipboard.service.ts` was 588 lines: serialising nodes and restoring
them, with five near-identical name pickers wedged between the two.

`clipboard/` splits it by direction -- `copy.ts` and `paste.ts` -- with
`types.ts` for the shapes and `names.ts` for the naming. The five pickers
differ only in what counts as "already taken" and how the numbered candidate
reads, so the search is written once and each one supplies those two things;
five copies of the same loop become five one-line calls.

Their suffix stays `_1` rather than the `(1)` that `helpers/uniqName`
produces for newly created objects. The two look interchangeable and are
not: the suffix shows up in the scene tree, and paste has always named this
way. `helpers/uniqName` stays where it is -- the surface and symmetry-copy
services use it.

No behaviour change; the 35 existing clipboard tests pass with only their
import path moved.
